Some cleanup and add support for a simplified az:// URI (#174)
The cleanup is to mainly aid the gatk logging of issues w.r.t to errors in the Cloud Storage systems.

Also add support for a simplified az URI of the type az://<container>/<path> especially if the env contains AZURE_ACCOUNT_NAME and the endpoint is the default blob.core.windows.net. This is in addition to az://<container>@<account_name>.<endpoint>/<path>. Makes the az and azb URIs identical if the account information is in env and/or the endpoint is the default.